+++ hash/tests/cshake_128_tester.c
@@ -64,9 +64,9 @@ static int _cshake_128_tester(const struct lc_hash *cshake_128,
 	LC_HASH_CTX_ON_STACK(ctx, cshake_128);
 	LC_CSHAKE_128_CTX_ON_STACK(cshake128_stack);
 
-	printf("hash ctx %s (%s implementation) len %lu\n", name,
+	printf("hash ctx %s (%s implementation) len %u\n", name,
 	       cshake_128 == lc_cshake128_c ? "C" : "accelerated",
-	       LC_HASH_CTX_SIZE(cshake_128));
+	       (unsigned int)LC_HASH_CTX_SIZE(cshake_128));
 
 	lc_cshake_init(ctx, NULL, 0, cust1, sizeof(cust1));
 	lc_hash_update(ctx, msg1, sizeof(msg1));
