+++ hash/tests/sha3-224_tester.c
@@ -47,9 +47,9 @@ static int _sha3_224_tester(const struct lc_hash *sha3_224, const char *name)
 	LC_HASH_CTX_ON_STACK(ctx224, sha3_224);
 	LC_SHA3_224_CTX_ON_STACK(ctx224_stack);
 
-	printf("hash ctx %s (%s implementation) len %lu\n", name,
+	printf("hash ctx %s (%s implementation) len %u\n", name,
 	       sha3_224 == lc_sha3_224_c ? "C" : "accelerated",
-	       LC_HASH_CTX_SIZE(sha3_224));
+	       (unsigned int)LC_HASH_CTX_SIZE(sha3_224));
 	lc_hash_init(ctx224);
 	lc_hash_update(ctx224, msg_224, 3);
 	lc_hash_final(ctx224, act);
