+++ hash/tests/sha3-512_tester.c
@@ -51,9 +51,9 @@ static int _sha3_512_tester(const struct lc_hash *sha3_512, const char *name)
 	LC_HASH_CTX_ON_STACK(ctx512, sha3_512);
 	LC_SHA3_512_CTX_ON_STACK(ctx512_stack);
 
-	printf("hash ctx %s (%s implementation) len %lu\n", name,
+	printf("hash ctx %s (%s implementation) len %u\n", name,
 	       sha3_512 == lc_sha3_512_c ? "C" : "accelerated",
-	       LC_HASH_CTX_SIZE(sha3_512));
+	       (unsigned int)LC_HASH_CTX_SIZE(sha3_512));
 	lc_hash_init(ctx512);
 	lc_hash_update(ctx512, msg_512, 3);
 	lc_hash_final(ctx512, act);
